+++ src/peer/media/sender/component.rs
@@ -127,7 +127,7 @@ pub struct State {
     /// Indicator whether this [`Sender`] is working in a [P2P mesh] or [SFU]
     /// mode.
     ///
-    /// [P2P mesh]: https://bloggeek.me/webrtcglossary/mesh
+    /// [P2P mesh]: https://webrtcglossary.com/mesh
     /// [SFU]: https://webrtcglossary.com/sfu
     connection_mode: ConnectionMode,